Uploaded "xml_to_json_converter.py"
Python script to convert an xml file to a json file. I was working with json files and wanted to explore how xml files worked. I thought a converter would be a good mini-project, as it allowed me to also gain a deeper understanding of working with json. I thought this converter would also have some actual value to me. ~5.5 hours of work.